+++ NEWS.md
@@ -1,5 +1,7 @@
 # recipes 0.1.9
 
+* `step_cut()` enables you to create a factor from a numeric based on provided break (contributed by Edwin Thoen)
+
  * Delayed S3 method registration for `tune::tunable()` methods that live in recipes will now work correctly on R >=4.0.0 ([#439](https://github.com/tidymodels/recipes/issues/439), [tidymodels/tune#146](https://github.com/tidymodels/tune/issues/146)).
  
  * `step_relevel()` added.